More cleanup of notifications
